fix: CVE should be sorted numeric
The CVE identifier contains letters and numbers. Using only the default
sorted function from python will not sort them correctly. The trick is
to give the key option a tuple which contains two values, the year and
the cve number. So it first will be sorted by year, then by number.

Fixes: #2